refactor(readDataSourceConfig): move prismaSqliteURLToFilePath util to separate folder, add supported providers validation, replace comments by js docs
